Moved the globalWaitlist.js file to the models directory
- Moved the `globalWaitlist.js` file to the models directory
- Removed the globalWaitlist from the `apiController.js`.
- Updated the reference for globalWaitlist in `fleetManagementController.js`, `pagesController.js`, `index.js`, and `fleets.js`
