+++ controllers/fleetManagementController.js
@@ -4,7 +4,7 @@ var fleets = require('../models/fleets.js')(setup);
 var users = require('../users.js')(setup);
 var api  = require('./apiController');
 var refresh = require('passport-oauth2-refresh');
-var waitlist = require('../globalWaitlist.js')(setup);
+var waitlist = require('../models/globalWaitlist.js')(setup);
 const log = require('../logger.js')(module);
 const wlog = require('../wlog.js');
